Removed unnecessary test files; added test script for parsing; ready for code review
