+++ qwik-dev-tool/src/application/algorithms/getHtmlTest.ts
@@ -1,0 +1,1538 @@
+const testHtml = `
+<html lang="en-us" q:container="paused" q:version="0.100.0" q:render="ssr-dev" q:base="/build/" q:locale=""><!--qv q:id=0 q:key=3scc:pY_0--><!--qv q:id=1 q:key=TxCF:35_3--><!--qv q:s q:sref=1 q:key=--><head q:head=""><meta charset="utf-8" q:head=""><link rel="manifest" href="/manifest.json" q:head=""><!--qv q:id=2 q:key=zrbr:35_0--><!--qv q:key=OA_0--><title q:head="">Qwik Todo List</title><link rel="canonical" href="http://localhost:5174/demo/todolist/" q:id="3" q:head=""><meta name="viewport" content="width=device-width, initial-scale=1.0" q:head=""><link rel="icon" type="image/svg+xml" href="/favicon.svg" q:head=""><!--/qv--><!--/qv--><link rel="stylesheet" href="/src/global.css"><link rel="stylesheet" href="/src/routes/demo/todolist/todolist.module.css"></head><body lang="en" data-qwik-inspector="root.tsx:22:7"><!--qv q:id=4 q:key=AKet:35_1--><!--qv q:id=5 q:key=VkLN:--><div class="page" q:key="XF_1" data-qwik-inspector="routes/layout.tsx:12:5"><main data-qwik-inspector="routes/layout.tsx:13:7"><!--qv q:s q:sref=5 q:key=--><!--qv q:id=6 q:key=J4V2:--><!--qv q:key=AP_3--><div class="section" data-qwik-inspector="routes/demo/todolist/index.tsx:40:7"><div class="container center" data-qwik-inspector="routes/demo/todolist/index.tsx:41:9"><h1 class="hero" data-qwik-inspector="routes/demo/todolist/index.tsx:42:11">TODO List</h1></div></div><div class="section bright" data-qwik-inspector="routes/demo/todolist/index.tsx:46:7"><div class="container center mh-300" data-qwik-inspector="routes/demo/todolist/index.tsx:47:9"><ul class="_list_14k68_1" q:key="AP_0" data-qwik-inspector="routes/demo/todolist/index.tsx:49:13"><li q:key="items-0" data-qwik-inspector="routes/demo/todolist/index.tsx:51:17">asd</li></ul></div></div><div class="section" data-qwik-inspector="routes/demo/todolist/index.tsx:58:7"><div class="container center" data-qwik-inspector="routes/demo/todolist/index.tsx:59:9"><form action="?qaction=tCQ608gnjnQ" preventdefault:submit="" method="post" data-spa-reset="true" on:submit="/src/routeactionqrl_action_submit_a5bzc7wo00a.js#routeActionQrl_action_submit_A5bZC7WO00A[0 1 2 3]" q:key="AP_2" q:id="7"><input type="text" name="text" required="" data-qwik-inspector="routes/demo/todolist/index.tsx:61:13"> <button type="submit" data-qwik-inspector="routes/demo/todolist/index.tsx:61:56">Add item</button></form><p class="_hint_14k68_11" data-qwik-inspector="routes/demo/todolist/index.tsx:64:11">PS: This little app works even when JavaScript is disabled.</p></div></div><!--/qv--><!--/qv--><!--/qv--></main></div><!--/qv--><script type="placeholder" hidden="" q:id="8" on-document:qinit="/src/routeroutlet_component_useondocument_event_knne9el0qfc.js#RouterOutlet_component_useOnDocument_event_KnNE9eL0qfc"></script><!--/qv--><!--qv q:key=35_2--><script>
+navigator.serviceWorker?.getRegistrations().then((regs) => {
+  for (const reg of regs) {
+    reg.unregister();
+  }
+});
+</script><!--/qv--><script type="qwik/json">{
+  "refs": {
+    "7": "6 q 1! 3!"
+  },
+  "ctx": {
+    "1": {
+      "w": "t",
+      "c": "qc-c=7! qc-ic=5 qc-h=0! qc-l=1! qc-n=j qc-s=4! qc-a=6"
+    },
+    "2": {
+      "h": "u",
+      "s": "0! 1!"
+    },
+    "4": {
+      "h": "8",
+      "s": "5"
+    },
+    "6": {
+      "h": "k",
+      "s": "o 1! 6 3!"
+    }
+  },
+  "objs": [
+    {
+      "title": "9",
+      "meta": "a",
+      "links": "b",
+      "styles": "c",
+      "frontmatter": "d"
+    },
+    {
+      "url": "f",
+      "params": "g",
+      "isNavigating": "h"
+    },
+    "\u0012i",
+    {
+      "actionPath": "p",
+      "isRunning": "h",
+      "status": "v",
+      "value": "v",
+      "formData": "v",
+      "submit": "r"
+    },
+    {
+      "VMJUTdq2hzw": "n"
+    },
+    "\u0012v",
+    "\u0012v",
+    {
+      "headings": "v",
+      "menu": "v"
+    },
+    "\u0002/src/routeroutlet_component_aketnbye5tm.js#RouterOutlet_component_AKetNByE5TM",
+    "Qwik Todo List",
+    [],
+    [],
+    [],
+    {},
+    "\u00111! @0",
+    "\u0005http://localhost:5174/demo/todolist/",
+    {},
+    false,
+    "/demo/todolist/",
+    "\u0002/src/qwikcityprovider_component_goto_fx0bdjeja0e.js#QwikCityProvider_component_goto_fX0bDjeJa0E[6 2 1!]",
+    "\u0002/src/todolist_component_j4v2qsf7yxo.js#todolist_component_J4V2qsF7Yxo",
+    "asd",
+    {
+      "text": "l"
+    },
+    [
+      "m"
+    ],
+    "\u00134! VMJUTdq2hzw",
+    "?qaction=tCQ608gnjnQ",
+    "tCQ608gnjnQ",
+    "\u0002/src/routeactionqrl_action_submit_a5bzc7wo00a.js#routeActionQrl_action_submit_A5bZC7WO00A[6 q 1! 3!]",
+    "\u0002/src/qwikcityprovider_component_usetask_02wmimzeabk.js#QwikCityProvider_component_useTask_02wMImzEAbk[6 7! 5 0! v 4! 2 1!]",
+    "\u00032 e s #1",
+    "\u0002/src/routerhead_component_zrbrqoaqxsy.js#RouterHead_component_zrbrqoaqXSY",
+    "\u0001"
+  ],
+  "subs": [
+    [
+      "0 #2 title",
+      "0 #2 meta",
+      "0 #2 links",
+      "0 #2 styles"
+    ],
+    [
+      "1 #3 e #2 href"
+    ],
+    [
+      "0 t"
+    ],
+    [
+      "0 #6 actionPath",
+      "0 #6 submit"
+    ],
+    [
+      "0 #6 VMJUTdq2hzw"
+    ],
+    [
+      "0 #4"
+    ],
+    [
+      "0 t"
+    ]
+  ]
+}</script><script q:func="qwik/json">document.currentScript.qFuncs=[(p0)=>(p0.url.href)]</script><script id="qwikloader">(() => {
+    ((doc, hasInitialized) => {
+        const win = window;
+        const events =  new Set;
+        const querySelectorAll = query => doc.querySelectorAll(query);
+        const broadcast = (infix, ev, type = ev.type) => {
+            querySelectorAll("[on" + infix + "\\:" + type + "]").forEach((target => dispatch(target, infix, ev, type)));
+        };
+        const getAttribute = (el, name) => el.getAttribute(name);
+        const resolveContainer = containerEl => {
+            if (void 0 === containerEl._qwikjson_) {
+                let script = (containerEl === doc.documentElement ? doc.body : containerEl).lastElementChild;
+                while (script) {
+                    if ("SCRIPT" === script.tagName && "qwik/json" === getAttribute(script, "type")) {
+                        containerEl._qwikjson_ = JSON.parse(script.textContent.replace(/\\x3C(\/?script)/g, "<$1"));
+                        break;
+                    }
+                    script = script.previousElementSibling;
+                }
+            }
+        };
+        const createEvent = (eventName, detail) => new CustomEvent(eventName, {
+            detail: detail
+        });
+        const dispatch = async (element, onPrefix, ev, eventName = ev.type) => {
+            const attrName = "on" + onPrefix + ":" + eventName;
+            element.hasAttribute("preventdefault:" + eventName) && ev.preventDefault();
+            const ctx = element._qc_;
+            const qrls = null == ctx ? void 0 : ctx.li.filter((li => li[0] === attrName));
+            if (qrls && qrls.length > 0) {
+                for (const q of qrls) {
+                    await q[1].getFn([ element, ev ], (() => element.isConnected))(ev, element);
+                }
+                return;
+            }
+            const attrValue = getAttribute(element, attrName);
+            if (attrValue) {
+                const container = element.closest("[q\\:container]");
+                const base = new URL(getAttribute(container, "q:base"), doc.baseURI);
+                for (const qrl of attrValue.split("\n")) {
+                    const url = new URL(qrl, base);
+                    const symbolName = url.hash.replace(/^#?([^?[|]*).*$/, "$1") || "default";
+                    const reqTime = performance.now();
+                    const module = import(url.href.split("#")[0]);
+                    resolveContainer(container);
+                    const handler = findSymbol(await module, symbolName);
+                    const previousCtx = doc.__q_context__;
+                    if (element.isConnected) {
+                        try {
+                            doc.__q_context__ = [ element, ev, url ];
+                            emitEvent("qsymbol", {
+                                symbol: symbolName,
+                                element: element,
+                                reqTime: reqTime
+                            });
+                            await handler(ev, element);
+                        } finally {
+                            doc.__q_context__ = previousCtx;
+                        }
+                    }
+                }
+            }
+        };
+        const emitEvent = (eventName, detail) => {
+            doc.dispatchEvent(createEvent(eventName, detail));
+        };
+        const findSymbol = (module, symbol) => {
+            if (symbol in module) {
+                return module[symbol];
+            }
+            for (const v of Object.values(module)) {
+                if ("object" == typeof v && v && symbol in v) {
+                    return v[symbol];
+                }
+            }
+        };
+        const camelToKebab = str => str.replace(/([A-Z])/g, (a => "-" + a.toLowerCase()));
+        const processDocumentEvent = async ev => {
+            let type = camelToKebab(ev.type);
+            let element = ev.target;
+            broadcast("-document", ev, type);
+            while (element && element.getAttribute) {
+                await dispatch(element, "", ev, type);
+                element = ev.bubbles && !0 !== ev.cancelBubble ? element.parentElement : null;
+            }
+        };
+        const processWindowEvent = ev => {
+            broadcast("-window", ev, camelToKebab(ev.type));
+        };
+        const processReadyStateChange = () => {
+            var _a;
+            const readyState = doc.readyState;
+            if (!hasInitialized && ("interactive" == readyState || "complete" == readyState)) {
+                hasInitialized = 1;
+                emitEvent("qinit");
+                (null != (_a = win.requestIdleCallback) ? _a : win.setTimeout).bind(win)((() => emitEvent("qidle")));
+                if (events.has("qvisible")) {
+                    const results = querySelectorAll("[on\\:qvisible]");
+                    const observer = new IntersectionObserver((entries => {
+                        for (const entry of entries) {
+                            if (entry.isIntersecting) {
+                                observer.unobserve(entry.target);
+                                dispatch(entry.target, "", createEvent("qvisible", entry));
+                            }
+                        }
+                    }));
+                    results.forEach((el => observer.observe(el)));
+                }
+            }
+        };
+        const addEventListener = (el, eventName, handler, capture = !1) => el.addEventListener(eventName, handler, {
+            capture: capture
+        });
+        const push = eventNames => {
+            for (const eventName of eventNames) {
+                if (!events.has(eventName)) {
+                    addEventListener(doc, eventName, processDocumentEvent, !0);
+                    addEventListener(win, eventName, processWindowEvent);
+                    events.add(eventName);
+                }
+            }
+        };
+        if (!doc.qR) {
+            const qwikevents = win.qwikevents;
+            Array.isArray(qwikevents) && push(qwikevents);
+            win.qwikevents = {
+                push: (...e) => push(e)
+            };
+            addEventListener(doc, "readystatechange", processReadyStateChange);
+            processReadyStateChange();
+        }
+    })(document);
+})();</script><script>window.qwikevents.push("submit", "qinit")</script>
+<style>
+vite-error-overlay {
+  --color-bright: rgba(255, 255, 255, 0.8);
+  --color-yellow: rgba(255,246,85,0.8);
+  --qwik-dark-blue: #006ce9;
+  --qwik-light-blue: #3ec2f7;
+  --qwik-light-purple: #ac7ff4;
+  --qwik-dark-purple: #713fc2;
+  --yellow: #fff;                   /* override vite yellow */
+  --purple: var(--color-bright);    /* override vite purple */
+  --red: var(--qwik-light-blue);    /* override vite red */
+
+  --vertical-box-spacing: 15px;
+  --box-padding: 20px;
+  --box-margin: 0 0 var(--vertical-box-spacing) 0;
+  --box-background: rgba(0, 0, 0, 0.5);
+  --box-border-radius: 8px;
+}
+
+vite-error-overlay::part(backdrop) {
+  background: rgb(2 11 17 / 60%);
+  backdrop-filter: blur(20px) brightness(0.4) saturate(3);
+}
+
+vite-error-overlay::part(window) {
+  background: transparent;
+  border: none;
+  box-shadow: none;
+  box-sizing: border-box;
+  margin: 50px auto;
+  max-width: 1200px;
+  padding: var(--box-padding);
+  width: 90%;
+}
+
+vite-error-overlay::part(message) {
+  display: flex;
+  flex-direction: column;
+  font-size: 1.6rem;
+  line-height: 1.7;
+  margin-bottom: 30px;
+}
+
+vite-error-overlay::part(plugin) {
+  font-size: 0.8rem;
+  font-weight: 100;
+}
+
+vite-error-overlay::part(file),
+vite-error-overlay::part(frame),
+vite-error-overlay::part(stack),
+vite-error-overlay::part(tip) {
+  background: var(--box-background);
+  border-left: 5px solid transparent;
+  border-radius: var(--box-border-radius);
+  margin: var(--box-margin);
+  min-height: 50px;
+  padding: var(--box-padding);
+  position: relative;
+}
+
+vite-error-overlay::part(file) {
+  border-left-color: rgb(25 182 246);
+  color: var(--color-bright);
+}
+
+vite-error-overlay::part(frame) {
+  border-left-color: var(--color-yellow);
+  color: var(--color-yellow);
+}
+
+vite-error-overlay::part(stack) {
+  border-left-color: #FF5722;
+}
+
+
+vite-error-overlay::part(tip) {
+  border-top: none;
+  border-left-color: rgb(172, 127, 244);
+}
+
+vite-error-overlay::part(file):before,
+vite-error-overlay::part(frame):before,
+vite-error-overlay::part(stack):before {
+  border-bottom: 1px solid #222;
+  color: var(--color-bright);
+  display: block;
+  margin-bottom: 15px;
+  padding-bottom: 5px;
+  padding-left: 30px; /* space for icon */
+  font-size: .8rem;
+}
+
+vite-error-overlay::part(file):before {
+  content: 'File';
+}
+
+vite-error-overlay::part(frame):before {
+  content: 'Frame';
+}
+
+vite-error-overlay::part(stack):before {
+  content: 'Stack Trace';
+}
+
+vite-error-overlay::part(file):after,
+vite-error-overlay::part(frame):after,
+vite-error-overlay::part(stack):after {
+  content: '';
+  display: block;
+  height: 20px;
+  position: absolute;
+  left: var(--box-padding);
+  top: var(--box-padding);
+  width: 20px;
+}
+
+vite-error-overlay::part(file):after {
+  background-image: url("data:image/svg+xml,%3Csvg width='20px' height='20px' viewBox='0 0 24 24' version='1.1' xmlns='http://www.w3.org/2000/svg' xmlns:xlink='http://www.w3.org/1999/xlink'%3E%3Ctitle%3EFile-Generic%3C/title%3E%3Cg id='Page-1' stroke='none' stroke-width='1' fill='none' fill-rule='evenodd'%3E%3Cg id='File-Generic'%3E%3Crect id='Rectangle' fill-rule='nonzero' x='0' y='0' width='24' height='24'%3E%3C/rect%3E%3Cpath d='M4 5 C4 3.89543 4.89543 3 6 3 L15.1716 3 C15.702 3 16.2107 3.21071 16.5858 3.58579 L19.4142 6.41421 C19.7893 6.78929 20 7.29799 20 7.82843 L20 19 C20 20.1046 19.1046 21 18 21 L6 21 C4.89543 21 4 20.1046 4 19 L4 5 Z' id='Path' stroke='rgba(255,255,255,0.7)' stroke-width='1' stroke-linecap='round'%3E%3C/path%3E%3Cpath d='M15 4 L15 6 C15 7.10457 15.8954 8 17 8 L19 8' id='Path' stroke='rgba(255,255,255,0.7)' stroke-width='1' stroke-linecap='round'%3E%3C/path%3E%3C/g%3E%3C/g%3E%3C/svg%3E");
+}
+
+vite-error-overlay::part(frame):after {
+  background-image: url("data:image/svg+xml,%3Csvg width='20px' height='20px' viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M15.6602 2.84952H19.1516C20.2555 2.84952 21.1504 3.74444 21.1504 4.84839V8.3398' stroke='rgba(255,255,255,0.7)' stroke-width='1.69904' stroke-linecap='round'/%3E%3Cpath d='M2.84949 8.33981L2.84949 4.8484C2.84949 3.74446 3.74441 2.84953 4.84836 2.84953L8.33977 2.84953' stroke='rgba(255,255,255,0.7)' stroke-width='1.69904' stroke-linecap='round'/%3E%3Cpath d='M21.1505 15.6602L21.1505 19.1516C21.1505 20.2555 20.2556 21.1505 19.1516 21.1505L15.6602 21.1505' stroke='rgba(255,255,255,0.7)' stroke-width='1.69904' stroke-linecap='round'/%3E%3Cpath d='M8.33984 21.1505L4.84843 21.1505C3.74449 21.1505 2.84956 20.2555 2.84956 19.1516L2.84956 15.6602' stroke='rgba(255,255,255,0.7)' stroke-width='1.69904' stroke-linecap='round'/%3E%3C/svg%3E");
+}
+
+vite-error-overlay::part(stack):after {
+  background-image: url("data:image/svg+xml,%3Csvg width='20px' height='20px' viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M14.78 20H9.78C7.98 20 4.58 19.09 4.58 15.64C4.58 12.19 7.98 11.28 9.78 11.28H14.22C14.37 11.28 17.92 11.23 17.92 8.42C17.92 5.61 14.37 5.56 14.22 5.56H9.22C9.02109 5.56 8.83032 5.48098 8.68967 5.34033C8.54902 5.19968 8.47 5.00891 8.47 4.81C8.47 4.61109 8.54902 4.42032 8.68967 4.27967C8.83032 4.13902 9.02109 4.06 9.22 4.06H14.22C16.02 4.06 19.42 4.97 19.42 8.42C19.42 11.87 16.02 12.78 14.22 12.78H9.78C9.63 12.78 6.08 12.83 6.08 15.64C6.08 18.45 9.63 18.5 9.78 18.5H14.78C14.9789 18.5 15.1697 18.579 15.3103 18.7197C15.451 18.8603 15.53 19.0511 15.53 19.25C15.53 19.4489 15.451 19.6397 15.3103 19.7803C15.1697 19.921 14.9789 20 14.78 20Z' fill='rgba(255,255,255,0.7)'/%3E%3Cpath d='M6.44 8.31C5.74314 8.30407 5.06363 8.09202 4.48708 7.70056C3.91054 7.30909 3.46276 6.75573 3.20018 6.11021C2.93759 5.46469 2.87195 4.75589 3.01153 4.07312C3.1511 3.39036 3.48965 2.76418 3.9845 2.2735C4.47935 1.78281 5.10837 1.44958 5.79229 1.31579C6.47622 1.182 7.18444 1.25363 7.82771 1.52167C8.47099 1.78971 9.02054 2.24215 9.40711 2.82199C9.79368 3.40182 9.99998 4.08311 10 4.78C10 5.2461 9.90773 5.70759 9.72846 6.13783C9.54919 6.56808 9.28648 6.95856 8.95551 7.28675C8.62453 7.61494 8.23184 7.87433 7.80009 8.04995C7.36834 8.22558 6.90609 8.31396 6.44 8.31ZM6.44 2.75C6.04444 2.75 5.65776 2.86729 5.32886 3.08706C4.99996 3.30682 4.74362 3.61918 4.59224 3.98463C4.44087 4.35008 4.40126 4.75221 4.47843 5.14018C4.5556 5.52814 4.74609 5.8845 5.02579 6.16421C5.3055 6.44391 5.66186 6.6344 6.04982 6.71157C6.43779 6.78874 6.83992 6.74913 7.20537 6.59776C7.57082 6.44638 7.88318 6.19003 8.10294 5.86114C8.32271 5.53224 8.44 5.14556 8.44 4.75C8.44 4.48735 8.38827 4.22728 8.28776 3.98463C8.18725 3.74198 8.03993 3.5215 7.85422 3.33578C7.6685 3.15007 7.44802 3.00275 7.20537 2.90224C6.96272 2.80173 6.70265 2.75 6.44 2.75Z' fill='rgba(255,255,255,0.7)'/%3E%3Cpath d='M17.56 22.75C16.8614 22.752 16.1779 22.5466 15.5961 22.1599C15.0143 21.7733 14.5603 21.2227 14.2916 20.5778C14.0229 19.933 13.9515 19.2229 14.0866 18.5375C14.2217 17.8521 14.5571 17.2221 15.0504 16.7275C15.5437 16.2328 16.1726 15.8956 16.8577 15.7586C17.5427 15.6215 18.253 15.6909 18.8986 15.9577C19.5442 16.2246 20.0961 16.6771 20.4844 17.2578C20.8727 17.8385 21.08 18.5214 21.08 19.22C21.08 20.1545 20.7095 21.0508 20.0496 21.7125C19.3898 22.3743 18.4945 22.7473 17.56 22.75ZM17.56 17.19C17.1644 17.19 16.7778 17.3073 16.4489 17.5271C16.12 17.7468 15.8636 18.0592 15.7122 18.4246C15.5609 18.7901 15.5213 19.1922 15.5984 19.5802C15.6756 19.9681 15.8661 20.3245 16.1458 20.6042C16.4255 20.8839 16.7819 21.0744 17.1698 21.1516C17.5578 21.2287 17.9599 21.1891 18.3254 21.0377C18.6908 20.8864 19.0032 20.63 19.2229 20.3011C19.4427 19.9722 19.56 19.5856 19.56 19.19C19.56 18.6596 19.3493 18.1508 18.9742 17.7758C18.5991 17.4007 18.0904 17.19 17.56 17.19Z' fill='rgba(255,255,255,0.7)'/%3E%3C/svg%3E");
+}
+
+vite-error-overlay::part(tip):before {
+  content: "Not sure how to solve this? Visit https://qwik.builder.io or connect with the community on Discord.";
+  display: block;
+  margin-bottom: 1em;
+}
+</style>
+<script type="module" src="/@vite/client"></script>
+
+<script>
+
+document.addEventListener('qerror', ev => {
+  const ErrorOverlay = customElements.get('vite-error-overlay');
+  if (!ErrorOverlay) {
+    return;
+  }
+  const err = ev.detail.error;
+  const overlay = new ErrorOverlay(err);
+  document.body.appendChild(overlay);
+});
+</script>
+
+<script>
+/**
+ * Usage:
+ *
+ *  <errored-host></errored-host>
+ *
+ */
+class ErroredHost extends HTMLElement {
+  get _root() {
+    return this.shadowRoot || this;
+  }
+
+  constructor() {
+    super();
+    const self = this;
+
+    this.state = {};
+    if (!this.props) {
+      this.props = {};
+    }
+
+    this.componentProps = ["children", "error"];
+
+    // used to keep track of all nodes created by show/for
+    this.nodesToDestroy = [];
+    // batch updates
+    this.pendingUpdate = false;
+
+    this.attachShadow({ mode: "open" });
+  }
+
+  destroyAnyNodes() {
+    // destroy current view template refs before rendering again
+    this.nodesToDestroy.forEach((el) => el.remove());
+    this.nodesToDestroy = [];
+  }
+
+  connectedCallback() {
+    this.getAttributeNames().forEach((attr) => {
+      const jsVar = attr.replace(/-/g, "");
+      const regexp = new RegExp(jsVar, "i");
+      this.componentProps.forEach((prop) => {
+        if (regexp.test(prop)) {
+          const attrValue = this.getAttribute(attr);
+          if (this.props[prop] !== attrValue) {
+            this.props[prop] = attrValue;
+          }
+        }
+      });
+    });
+
+    this._root.innerHTML = \`
+
+      <template data-el="show-errored-host">
+      <div class="error">
+        <template data-el="div-errored-host-2">
+          <!-- String(props.error) -->
+        </template>
+      </div>
+      </template>
+
+      <div class="arrow">👇 Uncaught error happened here 👇
+        <span class="dev-tools">DevTools: Cmd+Alt+I</span>
+      </div>
+      <div class="div">
+        <slot></slot>
+      </div>
+
+      <style>
+        .error {
+          border-radius: 5px 5px 0px 0;
+          background: black;
+          color: white;
+          font-family: monospace;
+          font-size: 12px;
+          margin: 0;
+          padding: 10px;
+        }
+        .arrow {
+          background: #f47e81;
+          color: black;
+          font-size: 14px;
+          padding: 10px;
+          text-align: center;
+          font-family: sans-serif;
+        }
+        .dev-tools {
+          background: red;
+          padding: 2px 5px;
+          border-radius: 3px;
+          font-weight: 800;
+        }
+        .div {
+          outline: 5px solid red;
+          border-radius: 10px;
+        }
+      </style>\`;
+    this.pendingUpdate = true;
+
+    this.render();
+    this.onMount();
+    this.pendingUpdate = false;
+    this.update();
+  }
+
+  showContent(el) {
+    // https://developer.mozilla.org/en-US/docs/Web/API/HTMLTemplateElement/content
+    // grabs the content of a node that is between <template> tags
+    // iterates through child nodes to register all content including text elements
+    // attaches the content after the template
+
+    const elementFragment = el.content.cloneNode(true);
+    const children = Array.from(elementFragment.childNodes);
+    children.forEach((child) => {
+      if (el?.scope) {
+        child.scope = el.scope;
+      }
+      if (el?.context) {
+        child.context = el.context;
+      }
+      this.nodesToDestroy.push(child);
+    });
+    el.after(elementFragment);
+  }
+
+  onMount() {}
+
+  onUpdate() {}
+
+  update() {
+    if (this.pendingUpdate === true) {
+      return;
+    }
+    this.pendingUpdate = true;
+    this.render();
+    this.onUpdate();
+    this.pendingUpdate = false;
+  }
+
+  render() {
+    // re-rendering needs to ensure that all nodes generated by for/show are refreshed
+    this.destroyAnyNodes();
+    this.updateBindings();
+  }
+
+  updateBindings() {
+    this._root
+      .querySelectorAll("[data-el='show-errored-host']")
+      .forEach((el) => {
+        const whenCondition = this.props.error;
+        if (whenCondition) {
+          this.showContent(el);
+        }
+      });
+
+    this._root
+      .querySelectorAll("[data-el='div-errored-host-2']")
+      .forEach((el) => {
+        this.renderTextNode(el, String(this.props.error));
+      });
+  }
+
+  // Helper to render content
+  renderTextNode(el, text) {
+    const textNode = document.createTextNode(text);
+    if (el?.scope) {
+      textNode.scope = el.scope;
+    }
+    if (el?.context) {
+      textNode.context = el.context;
+    }
+    el.after(textNode);
+    this.nodesToDestroy.push(el.nextSibling);
+  }
+}
+
+customElements.define("errored-host", ErroredHost);
+</script>
+
+
+<script>
+if (!window.__qwikViteLog) {
+  window.__qwikViteLog = true;
+  console.debug("%c⭐️ Qwik Dev SSR Mode","background: #0c75d2; color: white; padding: 2px 3px; border-radius: 2px; font-size: 0.8em;","App is running in SSR development mode!\n - Additional JS is loaded by Vite for debugging and live reloading\n - Rendering performance might not be optimal\n - Delayed interactivity because prefetching is disabled\n - Vite dev bundles do not represent production output\n\nProduction build can be tested running 'npm run preview'");
+}
+</script>
+
+<style>
+#qwik-inspector-overlay {
+  position: fixed;
+  background: rgba(24, 182, 246, 0.27);
+  pointer-events: none;
+  box-sizing: border-box;
+  border: 2px solid rgba(172, 126, 244, 0.46);
+  border-radius: 4px;
+  contain: strict;
+  cursor: pointer;
+  z-index: 999999;
+}
+#qwik-inspector-info-popup {
+  position: fixed;
+  bottom: 10px;
+  right: 10px;
+  font-family: monospace;
+  background: #000000c2;
+  color: white;
+  padding: 10px 20px;
+  border-radius: 8px;
+  box-shadow: 0 20px 25px -5px rgb(0 0 0 / 34%), 0 8px 10px -6px rgb(0 0 0 / 24%);
+  backdrop-filter: blur(4px);
+  -webkit-animation: fadeOut 0.3s 3s ease-in-out forwards;
+  animation: fadeOut 0.3s 3s ease-in-out forwards;
+  z-index: 999999;
+}
+#qwik-inspector-info-popup p {
+  margin: 0px;
+}
+@-webkit-keyframes fadeOut {
+  0% {opacity: 1;}
+  100% {opacity: 0;}
+}
+
+@keyframes fadeOut {
+  0% {opacity: 1;}
+  100% {opacity: 0; visibility: hidden;}
+}
+</style>
+<script>
+(function() {
+  console.debug("%c🔍 Qwik Click-To-Source","background: #564CE0; color: white; padding: 2px 3px; border-radius: 2px; font-size: 0.8em;","Hold-press the 'Alt' key and click a component to jump directly to the source code in your IDE!");
+  window.__qwik_inspector_state = {
+    pressedKeys: new Set(),
+  };
+  const origin = 'http://local.local';
+  const srcDir = new URL("/Users/michael-personal/git-repos/Codesmith/qwik-tutorial/app/src/", origin);
+  const body = document.body;
+  const overlay = document.createElement('div');
+  overlay.id = 'qwik-inspector-overlay';
+  overlay.setAttribute('aria-hidden', 'true');
+  body.appendChild(overlay);
+
+  document.addEventListener(
+    'keydown',
+    (event) => {
+      window.__qwik_inspector_state.pressedKeys.add(event.code);
+      updateOverlay();
+    },
+    { capture: true }
+  );
+
+  document.addEventListener(
+    'keyup',
+    (event) => {
+      window.__qwik_inspector_state.pressedKeys.delete(event.code);
+      updateOverlay();
+    },
+    { capture: true }
+  );
+
+  document.addEventListener(
+    'mouseover',
+    (event) => {
+      if (event.target && event.target instanceof HTMLElement && event.target.dataset.qwikInspector) {
+        window.__qwik_inspector_state.hoveredElement = event.target;
+      } else {
+        window.__qwik_inspector_state.hoveredElement = undefined;
+      }
+      updateOverlay();
+    },
+    { capture: true }
+  );
+
+  document.addEventListener(
+    'click',
+    (event) => {
+      if (isActive()) {
+        window.__qwik_inspector_state.pressedKeys.clear();
+        if (event.target && event.target instanceof HTMLElement) {
+          if (event.target.dataset.qwikInspector) {
+            event.preventDefault();
+            const resolvedURL = new URL(event.target.dataset.qwikInspector, srcDir);
+            body.style.setProperty('cursor', 'progress');
+            if (resolvedURL.origin === origin) {
+              const params = new URLSearchParams();
+              params.set('file', resolvedURL.pathname);
+              fetch('/__open-in-editor?' + params.toString());
+            } else {
+              location.href = resolvedURL.href;
+            }
+          }
+        }
+      }
+    },
+    { capture: true }
+  );
+
+  document.addEventListener(
+    'contextmenu',
+    (event) => {
+      if (isActive()) {
+        window.__qwik_inspector_state.pressedKeys.clear();
+        if (event.target && event.target instanceof HTMLElement) {
+          if (event.target.dataset.qwikInspector) {
+            event.preventDefault();
+          }
+        }
+      }
+    },
+    { capture: true }
+  );
+
+  function updateOverlay() {
+    const hoverElement = window.__qwik_inspector_state.hoveredElement;
+    if (hoverElement && isActive()) {
+      const rect = hoverElement.getBoundingClientRect();
+      overlay.style.setProperty('height', rect.height + 'px');
+      overlay.style.setProperty('width', rect.width + 'px');
+      overlay.style.setProperty('top', rect.top + 'px');
+      overlay.style.setProperty('left', rect.left + 'px');
+      overlay.style.setProperty('visibility', 'visible');
+      body.style.setProperty('cursor', 'pointer');
+    } else {
+      overlay.style.setProperty('height', '0px');
+      overlay.style.setProperty('width', '0px');
+      overlay.style.setProperty('visibility', 'hidden');
+      body.style.removeProperty('cursor');
+    }
+  }
+
+  function checkKeysArePressed() {
+    const activeKeys = Array.from(window.__qwik_inspector_state.pressedKeys)
+      .map((key) => key ? key.replace(/(Left|Right)$/g, '') : undefined);
+    const clickToSourceKeys = ["Alt"];
+    return clickToSourceKeys.every((key) => activeKeys.includes(key));
+  }
+
+  function isActive() {
+    return checkKeysArePressed();
+  }
+
+  window.addEventListener('resize', updateOverlay);
+  document.addEventListener('scroll', updateOverlay);
+
+})();
+</script><div id="qwik-inspector-overlay" aria-hidden="true" style="height: 0px; width: 0px; visibility: hidden;"></div>
+<div id="qwik-inspector-info-popup" aria-hidden="true">Click-to-Source: Alt<p></p></div>
+
+</body><!--/qv--><!--/qv--><!--/qv--></html>
+`
+
+
+const testHtml2 = `
+<html lang="en-us" q:container="paused" q:version="0.101.0" q:render="ssr-dev" q:base="/build/" q:locale=""><!--qv q:id=0 q:key=3scc:pY_0--><!--qv q:id=1 q:key=TxCF:35_3--><!--qv q:s q:sref=1 q:key=--><head q:head=""><meta charset="utf-8" q:head=""><link rel="manifest" href="/manifest.json" q:head=""><!--qv q:id=2 q:key=zrbr:35_0--><!--qv q:key=OA_0--><title q:head="">Welcome to Qwik</title><link rel="canonical" href="http://localhost:5174/" q:id="3" q:head=""><meta name="viewport" content="width=device-width, initial-scale=1.0" q:head=""><link rel="icon" type="image/svg+xml" href="/favicon.svg" q:head=""><meta name="description" content="Qwik site description" q:head=""><!--/qv--><!--/qv--><link rel="stylesheet" href="/src/components/starter/header/header.module.css"><link rel="stylesheet" href="/src/components/starter/footer/footer.module.css"><link rel="stylesheet" href="/src/components/starter/hero/hero.module.css"><link rel="stylesheet" href="/src/components/starter/infobox/infobox.module.css"><link rel="stylesheet" href="/src/components/starter/next-steps/next-steps.module.css"><link rel="stylesheet" href="/src/global.css"><link rel="stylesheet" href="/src/routes/demo/todolist/todolist.module.css"></head><body lang="en" data-qwik-inspector="root.tsx:22:7"><!--qv q:id=4 q:key=AKet:35_1--><!--qv q:id=5 q:key=VkLN:k8_0--><div class="page" q:key="XF_3" data-qwik-inspector="routes/layout.tsx:15:5"><main data-qwik-inspector="routes/layout.tsx:16:7"><!--qv q:id=6 q:key=mBt9:XF_0--><header class="_header_1lpd3_1" q:key="CP_1" data-qwik-inspector="components/starter/header/header.tsx:7:5"><div class="_logo_1lpd3_7" data-qwik-inspector="components/starter/header/header.tsx:8:7"><a href="/" title="qwik" data-qwik-inspector="components/starter/header/header.tsx:9:9"><!--qv q:key=CP_0--><svg width="100" height="35" viewBox="0 0 167 53" fill="none" xmlns="http://www.w3.org/2000/svg" q:key="Ph_0" data-qwik-inspector="components/starter/icons/qwik.tsx:2:3"><path d="M81.9545 46.5859H75.5513V35.4045C73.4363 36.8579 71.0496 37.5749 68.4884 37.5749C65.0151 37.5749 62.4344 36.6253 60.8239 34.6487C59.2134 32.6915 58.3984 29.2034 58.3984 24.2231C58.3984 19.1266 59.3492 15.5997 61.2702 13.5456C63.23 11.4721 66.3734 10.4644 70.7004 10.4644C74.7946 10.4644 78.5201 11.0264 81.9545 12.131V46.5859ZM75.5513 16.278C74.096 15.8323 72.4661 15.6191 70.7004 15.6191C68.5272 15.6191 66.9749 16.1811 66.1017 17.3244C65.2479 18.4871 64.7823 20.6962 64.7823 23.9712C64.7823 27.0524 65.1897 29.1065 66.0435 30.2304C66.8973 31.335 68.3719 31.897 70.5452 31.897C73.3781 31.897 75.5513 30.7343 75.5513 29.2809V16.278Z" fill="white" data-qwik-inspector="components/starter/icons/qwik.tsx:3:5"></path><path d="M91.133 11.1426C93.4033 17.4406 95.3242 23.7386 96.993 30.0948C99.205 23.5836 101.087 17.2856 102.542 11.1426H108.15C110.265 17.4406 112.031 23.7386 113.447 30.0948C115.97 23.196 117.949 16.8787 119.404 11.1426H125.71C123.033 20.173 120.064 28.777 116.785 36.8966H109.256C108.402 32.3039 107.044 26.7617 105.22 20.1536C104.056 25.2889 102.445 30.8893 100.33 36.8966H92.8018C90.2793 27.5174 87.5434 18.9522 84.6328 11.1426H91.133Z" fill="white" data-qwik-inspector="components/starter/icons/qwik.tsx:7:5"></path><path d="M132.832 7.55758C129.999 7.55758 129.203 6.85996 129.203 3.97257C129.203 1.39523 130.018 0.794495 132.832 0.794495C135.665 0.794495 136.46 1.39523 136.46 3.97257C136.46 6.85996 135.665 7.55758 132.832 7.55758ZM129.649 11.1426H136.053V36.8966H129.649V11.1426Z" fill="white" data-qwik-inspector="components/starter/icons/qwik.tsx:11:5"></path><path d="M166.303 11.1426C161.763 17.5956 158.581 21.5295 156.815 22.9441C158.27 23.8937 162.17 28.8933 167.002 36.916H159.628C153.613 27.7887 150.742 23.8549 149.325 23.2542V36.916H142.922V0H149.325V23.2348C150.78 22.169 153.963 18.1382 158.872 11.1426H166.303Z" fill="white" data-qwik-inspector="components/starter/icons/qwik.tsx:15:5"></path><path d="M40.973 52.5351L32.0861 43.6985L31.9503 43.7179V43.621L13.0511 24.9595L17.708 20.4637L14.9721 4.76715L1.99103 20.8513C-0.220992 23.0798 -0.628467 26.7036 0.962635 29.3778L9.07337 42.8265C10.3152 44.9 12.566 46.1402 14.9915 46.1208L19.0081 46.082L40.973 52.5351Z" fill="#18B6F6" data-qwik-inspector="components/starter/icons/qwik.tsx:19:5"></path><path d="M45.8232 20.5411L44.038 17.2468L43.1066 15.5609L42.738 14.902L42.6992 14.9408L37.8094 6.47238C36.587 4.34075 34.2974 3.02301 31.8137 3.04239L27.5255 3.15865L14.7384 3.19741C12.313 3.21679 10.101 4.49577 8.87853 6.56927L1.09766 21.9945L15.0101 4.72831L33.2496 24.7656L30.0091 28.0406L31.9495 43.7178L31.9689 43.679V43.7178H31.9301L31.9689 43.7565L33.4824 45.2293L40.8364 52.4187C41.1469 52.7094 41.6514 52.3606 41.4379 51.9924L36.8975 43.0589L44.8142 28.4282L45.0664 28.1375C45.1634 28.0212 45.2604 27.905 45.3381 27.7887C46.8904 25.6764 47.1038 22.8472 45.8232 20.5411Z" fill="#AC7EF4" data-qwik-inspector="components/starter/icons/qwik.tsx:23:5"></path><path d="M33.3076 24.6882L15.0099 4.74774L17.61 20.3668L12.9531 24.882L31.9105 43.6985L30.203 28.0794L33.3076 24.6882Z" fill="white" data-qwik-inspector="components/starter/icons/qwik.tsx:27:5"></path></svg><!--/qv--></a></div><ul data-qwik-inspector="components/starter/header/header.tsx:13:7"><li data-qwik-inspector="components/starter/header/header.tsx:14:9"><a href="https://qwik.builder.io/docs/components/overview/" target="_blank" data-qwik-inspector="components/starter/header/header.tsx:15:11">Docs</a></li><li data-qwik-inspector="components/starter/header/header.tsx:19:9"><a href="https://qwik.builder.io/examples/introduction/hello-world/" target="_blank" data-qwik-inspector="components/starter/header/header.tsx:20:11">Examples</a></li><li data-qwik-inspector="components/starter/header/header.tsx:24:9"><a href="https://qwik.builder.io/tutorial/welcome/overview/" target="_blank" data-qwik-inspector="components/starter/header/header.tsx:25:11">Tutorials</a></li></ul></header><!--/qv--><!--qv q:s q:sref=5 q:key=--><!--qv q:id=7 q:key=xYL1:k8_0--><!--qv q:key=H1_7--><!--qv q:id=8 q:key=fle1:H1_0--><div class="_hero_1xbdm_1" q:key="nG_0" data-qwik-inspector="components/starter/hero/hero.tsx:6:5"><h1 data-qwik-inspector="components/starter/hero/hero.tsx:7:7">Welcome to qwik</h1><button on:click="/src/hero_component_div_button_onclick_j3qmou80s0i.js#hero_component_div_button_onClick_j3QmOu80S0I" data-qwik-inspector="components/starter/hero/hero.tsx:8:7">Time to celebrate 🎉</button></div><!--/qv--><div class="section bright" data-qwik-inspector="routes/index.tsx:14:7"><div class="container center" data-qwik-inspector="routes/index.tsx:15:9"><!--qv q:id=9 q:key=kJCt:H1_1--><!--qv q:key=W5_1--><div class="_gettingstarted_1knif_1" on-window:keydown="/src/next_steps_component_useonwindow_uxljfslpf0s.js#next_steps_component_useOnWindow_UxlJFslpf0s[0]" q:id="a" data-qwik-inspector="components/starter/next-steps/next-steps.tsx:45:7"><div data-qwik-inspector="components/starter/next-steps/next-steps.tsx:46:9"><b>Ready</b> to make some changes?<br>Press and hold the <b>ALT</b> key</div><span class="_hint_1knif_10" data-qwik-inspector="components/starter/next-steps/next-steps.tsx:47:9"></span></div><button class="gray small" on:click="/src/next_steps_component__fragment_button_onclick_v2jle7d93co.js#next_steps_component__Fragment_button_onClick_v2JLE7D93co[0]" q:key="W5_0" q:id="b" data-qwik-inspector="components/starter/next-steps/next-steps.tsx:53:9">Continue with Step <!--t=c-->2<!----> of 6</button><!--/qv--><!--/qv--></div></div><div class="section" data-qwik-inspector="routes/index.tsx:20:7"><div class="container center" data-qwik-inspector="routes/index.tsx:21:9"><h3 data-qwik-inspector="routes/index.tsx:22:11">You can <b data-qwik-inspector="routes/index.tsx:23:21">count</b> on me</h3><!--qv q:sstyle=⭐️5vd6y7-0 q:id=d q:key=5Go3:H1_2--><style q:style="5vd6y7-0" hidden="">.counter-wrapper.⭐️5vd6y7-0 {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 15px;
+}
+
+.counter-value.⭐️5vd6y7-0 {
+  background: white;
+  border-radius: 12px;
+  color: var(--qwik-light-purple);
+  font-weight: 400;
+  font-size: 2rem;
+  width: 120px;
+}
+
+.counter-value.odd.⭐️5vd6y7-0 {
+  color: var(--qwik-light-blue);
+}
+</style><div class="⭐️5vd6y7-0 counter-wrapper" q:key="no_0" data-qwik-inspector="components/starter/counter/counter.tsx:9:5"><button class="⭐️5vd6y7-0" on:click="/src/counter_component_div_button_onclick_d04jayucnhm.js#counter_component_div_button_onClick_D04jAYuCnhM[0]" q:id="e" data-qwik-inspector="components/starter/counter/counter.tsx:10:7">-</button><span class="⭐️5vd6y7-0 counter-value odd" q:id="f" data-qwik-inspector="components/starter/counter/counter.tsx:11:7"><!--t=g-->10<!----></span><button class="⭐️5vd6y7-0" on:click="/src/counter_component_div_button_onclick_1_lkcvrojx09y.js#counter_component_div_button_onClick_1_LkCVrojX09Y[0]" q:id="h" data-qwik-inspector="components/starter/counter/counter.tsx:12:7">+</button></div><!--/qv--></div></div><div class="section" data-qwik-inspector="routes/index.tsx:29:7"><div class="container topics" data-qwik-inspector="routes/index.tsx:30:9"><!--qv q:id=i q:key=foRS:H1_4--><div class="_infobox_otv38_1" q:key="aD_2" data-qwik-inspector="components/starter/infobox/infobox.tsx:6:5"><h3 data-qwik-inspector="components/starter/infobox/infobox.tsx:7:7"><!--qv q:s q:sref=i q:key=title--><div q:slot="title" class="icon icon-cli" data-qwik-inspector="routes/index.tsx:32:13">CLI Commands</div><!--/qv--></h3><!--qv q:s q:sref=i q:key=--><p data-qwik-inspector="routes/index.tsx:36:15"><code data-qwik-inspector="routes/index.tsx:37:17">npm run dev</code><br data-qwik-inspector="routes/index.tsx:38:17">Starts the development server and watches for changes</p><p data-qwik-inspector="routes/index.tsx:41:15"><code data-qwik-inspector="routes/index.tsx:42:17">npm run preview</code><br data-qwik-inspector="routes/index.tsx:43:17">Creates production build and starts a server to preview it</p><p data-qwik-inspector="routes/index.tsx:46:15"><code data-qwik-inspector="routes/index.tsx:47:17">npm run build</code><br data-qwik-inspector="routes/index.tsx:48:17">Creates production build</p><p data-qwik-inspector="routes/index.tsx:51:15"><code data-qwik-inspector="routes/index.tsx:52:17">npm run qwik add</code><br data-qwik-inspector="routes/index.tsx:53:17">Runs the qwik CLI to add integrations</p><!--/qv--></div><!--/qv--><div data-qwik-inspector="routes/index.tsx:59:11"><!--qv q:id=j q:key=foRS:H1_5--><div class="_infobox_otv38_1" q:key="aD_2" data-qwik-inspector="components/starter/infobox/infobox.tsx:6:5"><h3 data-qwik-inspector="components/starter/infobox/infobox.tsx:7:7"><!--qv q:s q:sref=j q:key=title--><div q:slot="title" class="icon icon-apps" data-qwik-inspector="routes/index.tsx:61:15">Example Apps</div><!--/qv--></h3><!--qv q:s q:sref=j q:key=--><p data-qwik-inspector="routes/index.tsx:64:15">Have a look at the <a href="/demo/flower" data-qwik-inspector="routes/index.tsx:65:36">Flower App</a> or the <a href="/demo/todolist" data-qwik-inspector="routes/index.tsx:66:17">Todo App</a>.</p><!--/qv--></div><!--/qv--><!--qv q:id=k q:key=foRS:H1_6--><div class="_infobox_otv38_1" q:key="aD_2" data-qwik-inspector="components/starter/infobox/infobox.tsx:6:5"><h3 data-qwik-inspector="components/starter/infobox/infobox.tsx:7:7"><!--qv q:s q:sref=k q:key=title--><div q:slot="title" class="icon icon-community" data-qwik-inspector="routes/index.tsx:71:15">Community</div><!--/qv--></h3><!--qv q:s q:sref=k q:key=--><ul data-qwik-inspector="routes/index.tsx:74:15"><li data-qwik-inspector="routes/index.tsx:75:17"><span data-qwik-inspector="routes/index.tsx:76:19">Questions or just want to say hi? </span><a href="https://qwik.builder.io/chat" target="_blank" data-qwik-inspector="routes/index.tsx:77:19">Chat on discord!</a></li><li data-qwik-inspector="routes/index.tsx:81:17"><span data-qwik-inspector="routes/index.tsx:82:19">Follow </span><a href="https://twitter.com/QwikDev" target="_blank" data-qwik-inspector="routes/index.tsx:83:19">@QwikDev</a><span data-qwik-inspector="routes/index.tsx:86:19"> on Twitter</span></li><li data-qwik-inspector="routes/index.tsx:88:17"><span data-qwik-inspector="routes/index.tsx:89:19">Open issues and contribute on </span><a href="https://github.com/BuilderIO/qwik" target="_blank" data-qwik-inspector="routes/index.tsx:90:19">GitHub</a></li><li data-qwik-inspector="routes/index.tsx:94:17"><span data-qwik-inspector="routes/index.tsx:95:19">Watch </span><a href="https://qwik.builder.io/media/" target="_blank" data-qwik-inspector="routes/index.tsx:96:19">Presentations, Podcasts, Videos, etc.</a></li></ul><!--/qv--></div><!--/qv--></div></div></div><!--/qv--><!--/qv--><!--/qv--></main><div class="section dark" data-qwik-inspector="routes/layout.tsx:20:7"><div class="container" data-qwik-inspector="routes/layout.tsx:21:9"><!--qv q:id=l q:key=u0bw:XF_2--><footer q:key="so_0" data-qwik-inspector="components/starter/footer/footer.tsx:9:5"><a href="https://www.builder.io/" target="_blank" class="_anchor_vbcl2_1" data-qwik-inspector="components/starter/footer/footer.tsx:10:7">Made with ♡ by Builder.io<span class="_spacer_vbcl2_9" data-qwik-inspector="components/starter/footer/footer.tsx:12:9">|</span><span data-qwik-inspector="components/starter/footer/footer.tsx:13:9"><!--t=m-->2023-04-17T20:27:08.557Z<!----></span></a></footer><!--/qv--></div></div></div><!--/qv--><script type="placeholder" hidden="" q:id="n" on-document:qinit="/src/routeroutlet_component_useondocument_event_knne9el0qfc.js#RouterOutlet_component_useOnDocument_event_KnNE9eL0qfc"></script><!--/qv--><!--qv q:key=35_2--><script q:key="1Z_0">
+navigator.serviceWorker?.getRegistrations().then((regs) => {
+  for (const reg of regs) {
+    reg.unregister();
+  }
+});
+</script><!--/qv--><script type="qwik/json">{
+  "refs": {
+    "a": "0",
+    "b": "0",
+    "e": "1",
+    "h": "1"
+  },
+  "ctx": {
+    "1": {
+      "w": "10",
+      "c": "qc-c=f! qc-ic=7 qc-h=2! qc-l=3! qc-n=11 qc-s=4! qc-a=6"
+    },
+    "2": {
+      "h": "12",
+      "s": "2! 3!"
+    },
+    "4": {
+      "h": "g",
+      "s": "7"
+    },
+    "9": {
+      "h": "e",
+      "s": "0"
+    }
+  },
+  "objs": [
+    "\u00128",
+    "\u0012a",
+    {
+      "title": "h",
+      "meta": "l",
+      "links": "m",
+      "styles": "n",
+      "frontmatter": "o"
+    },
+    {
+      "url": "q",
+      "params": "r",
+      "isNavigating": "s"
+    },
+    {
+      "GQamrjryd1Y": "u"
+    },
+    "\u0012y",
+    "\u001213",
+    "\u001213",
+    0,
+    "\u00110 @0",
+    10,
+    "\u00111 @1",
+    "\u00111 @2",
+    "#g",
+    "\u0002/src/next_steps_component_kjctkbc9zbk.js#next_steps_component_kJCtKbc9zbk",
+    {
+      "headings": "13",
+      "menu": "13"
+    },
+    "\u0002/src/routeroutlet_component_aketnbye5tm.js#RouterOutlet_component_AKetNByE5TM",
+    "Welcome to Qwik",
+    "description",
+    "Qwik site description",
+    {
+      "name": "i",
+      "content": "j"
+    },
+    [
+      "k"
+    ],
+    [],
+    [],
+    {},
+    "\u00113! @3",
+    "\u0005http://localhost:5174/",
+    {},
+    false,
+    "2023-04-17T20:27:08.557Z",
+    {
+      "date": "t"
+    },
+    "\u00134! GQamrjryd1Y",
+    "\u0011v @4",
+    "#m",
+    "/",
+    "\u0002/src/qwikcityprovider_component_usetask_02wmimzeabk.js#QwikCityProvider_component_useTask_02wMImzEAbk[6 f! 7 2! 13 4! 5 3!]",
+    "\u00032 e z #1",
+    "\u0002/src/qwikcityprovider_component_goto_fx0bdjeja0e.js#QwikCityProvider_component_goto_fX0bDjeJa0E[6 5 3!]",
+    "\u0002/src/routerhead_component_zrbrqoaqxsy.js#RouterHead_component_zrbrqoaqXSY",
+    "\u0001"
+  ],
+  "subs": [
+    [
+      "0 #9",
+      "4 #9 9 #c"
+    ],
+    [
+      "1 #f b #d class",
+      "3 #g c #g"
+    ],
+    [
+      "0 #2 title",
+      "0 #2 meta",
+      "0 #2 links",
+      "0 #2 styles"
+    ],
+    [
+      "1 #3 p #2 href"
+    ],
+    [
+      "3 #m w #m"
+    ],
+    [
+      "0 10"
+    ],
+    [
+      "0 10"
+    ],
+    [
+      "0 #4"
+    ]
+  ]
+}</script><script q:func="qwik/json">document.currentScript.qFuncs=[(p0)=>(p0.value+2),
+(p0)=>(\`counter-value $\{p0.value%2===0?"odd":""}\`),
+(p0)=>(p0.value),
+(p0)=>(p0.url.href),
+(p0)=>(p0.value.date)]</script><script id="qwikloader">(() => {
+    ((doc, hasInitialized) => {
+        const win = window;
+        const events =  new Set;
+        const querySelectorAll = query => doc.querySelectorAll(query);
+        const broadcast = (infix, ev, type = ev.type) => {
+            querySelectorAll("[on" + infix + "\\:" + type + "]").forEach((target => dispatch(target, infix, ev, type)));
+        };
+        const getAttribute = (el, name) => el.getAttribute(name);
+        const resolveContainer = containerEl => {
+            if (void 0 === containerEl._qwikjson_) {
+                let script = (containerEl === doc.documentElement ? doc.body : containerEl).lastElementChild;
+                while (script) {
+                    if ("SCRIPT" === script.tagName && "qwik/json" === getAttribute(script, "type")) {
+                        containerEl._qwikjson_ = JSON.parse(script.textContent.replace(/\\x3C(\/?script)/g, "<$1"));
+                        break;
+                    }
+                    script = script.previousElementSibling;
+                }
+            }
+        };
+        const createEvent = (eventName, detail) => new CustomEvent(eventName, {
+            detail: detail
+        });
+        const dispatch = async (element, onPrefix, ev, eventName = ev.type) => {
+            const attrName = "on" + onPrefix + ":" + eventName;
+            element.hasAttribute("preventdefault:" + eventName) && ev.preventDefault();
+            const ctx = element._qc_;
+            const qrls = null == ctx ? void 0 : ctx.li.filter((li => li[0] === attrName));
+            if (qrls && qrls.length > 0) {
+                for (const q of qrls) {
+                    await q[1].getFn([ element, ev ], (() => element.isConnected))(ev, element);
+                }
+                return;
+            }
+            const attrValue = getAttribute(element, attrName);
+            if (attrValue) {
+                const container = element.closest("[q\\:container]");
+                const base = new URL(getAttribute(container, "q:base"), doc.baseURI);
+                for (const qrl of attrValue.split("\n")) {
+                    const url = new URL(qrl, base);
+                    const symbolName = url.hash.replace(/^#?([^?[|]*).*$/, "$1") || "default";
+                    const reqTime = performance.now();
+                    const module = import(url.href.split("#")[0]);
+                    resolveContainer(container);
+                    const handler = findSymbol(await module, symbolName);
+                    const previousCtx = doc.__q_context__;
+                    if (element.isConnected) {
+                        try {
+                            doc.__q_context__ = [ element, ev, url ];
+                            emitEvent("qsymbol", {
+                                symbol: symbolName,
+                                element: element,
+                                reqTime: reqTime
+                            });
+                            await handler(ev, element);
+                        } finally {
+                            doc.__q_context__ = previousCtx;
+                        }
+                    }
+                }
+            }
+        };
+        const emitEvent = (eventName, detail) => {
+            doc.dispatchEvent(createEvent(eventName, detail));
+        };
+        const findSymbol = (module, symbol) => {
+            if (symbol in module) {
+                return module[symbol];
+            }
+            for (const v of Object.values(module)) {
+                if ("object" == typeof v && v && symbol in v) {
+                    return v[symbol];
+                }
+            }
+        };
+        const camelToKebab = str => str.replace(/([A-Z])/g, (a => "-" + a.toLowerCase()));
+        const processDocumentEvent = async ev => {
+            let type = camelToKebab(ev.type);
+            let element = ev.target;
+            broadcast("-document", ev, type);
+            while (element && element.getAttribute) {
+                await dispatch(element, "", ev, type);
+                element = ev.bubbles && !0 !== ev.cancelBubble ? element.parentElement : null;
+            }
+        };
+        const processWindowEvent = ev => {
+            broadcast("-window", ev, camelToKebab(ev.type));
+        };
+        const processReadyStateChange = () => {
+            var _a;
+            const readyState = doc.readyState;
+            if (!hasInitialized && ("interactive" == readyState || "complete" == readyState)) {
+                hasInitialized = 1;
+                emitEvent("qinit");
+                (null != (_a = win.requestIdleCallback) ? _a : win.setTimeout).bind(win)((() => emitEvent("qidle")));
+                if (events.has("qvisible")) {
+                    const results = querySelectorAll("[on\\:qvisible]");
+                    const observer = new IntersectionObserver((entries => {
+                        for (const entry of entries) {
+                            if (entry.isIntersecting) {
+                                observer.unobserve(entry.target);
+                                dispatch(entry.target, "", createEvent("qvisible", entry));
+                            }
+                        }
+                    }));
+                    results.forEach((el => observer.observe(el)));
+                }
+            }
+        };
+        const addEventListener = (el, eventName, handler, capture = !1) => el.addEventListener(eventName, handler, {
+            capture: capture
+        });
+        const push = eventNames => {
+            for (const eventName of eventNames) {
+                if (!events.has(eventName)) {
+                    addEventListener(doc, eventName, processDocumentEvent, !0);
+                    addEventListener(win, eventName, processWindowEvent);
+                    events.add(eventName);
+                }
+            }
+        };
+        if (!doc.qR) {
+            const qwikevents = win.qwikevents;
+            Array.isArray(qwikevents) && push(qwikevents);
+            win.qwikevents = {
+                push: (...e) => push(e)
+            };
+            addEventListener(doc, "readystatechange", processReadyStateChange);
+            processReadyStateChange();
+        }
+    })(document);
+})();</script><script>window.qwikevents.push("click", "keydown", "qinit")</script>
+<style>
+vite-error-overlay {
+  --color-bright: rgba(255, 255, 255, 0.8);
+  --color-yellow: rgba(255,246,85,0.8);
+  --qwik-dark-blue: #006ce9;
+  --qwik-light-blue: #3ec2f7;
+  --qwik-light-purple: #ac7ff4;
+  --qwik-dark-purple: #713fc2;
+  --yellow: #fff;                   /* override vite yellow */
+  --purple: var(--color-bright);    /* override vite purple */
+  --red: var(--qwik-light-blue);    /* override vite red */
+
+  --vertical-box-spacing: 15px;
+  --box-padding: 20px;
+  --box-margin: 0 0 var(--vertical-box-spacing) 0;
+  --box-background: rgba(0, 0, 0, 0.5);
+  --box-border-radius: 8px;
+}
+
+vite-error-overlay::part(backdrop) {
+  background: rgb(2 11 17 / 60%);
+  backdrop-filter: blur(20px) brightness(0.4) saturate(3);
+}
+
+vite-error-overlay::part(window) {
+  background: transparent;
+  border: none;
+  box-shadow: none;
+  box-sizing: border-box;
+  margin: 50px auto;
+  max-width: 1200px;
+  padding: var(--box-padding);
+  width: 90%;
+}
+
+vite-error-overlay::part(message) {
+  display: flex;
+  flex-direction: column;
+  font-size: 1.6rem;
+  line-height: 1.7;
+  margin-bottom: 30px;
+}
+
+vite-error-overlay::part(plugin) {
+  font-size: 0.8rem;
+  font-weight: 100;
+}
+
+vite-error-overlay::part(file),
+vite-error-overlay::part(frame),
+vite-error-overlay::part(stack),
+vite-error-overlay::part(tip) {
+  background: var(--box-background);
+  border-left: 5px solid transparent;
+  border-radius: var(--box-border-radius);
+  margin: var(--box-margin);
+  min-height: 50px;
+  padding: var(--box-padding);
+  position: relative;
+}
+
+vite-error-overlay::part(file) {
+  border-left-color: rgb(25 182 246);
+  color: var(--color-bright);
+}
+
+vite-error-overlay::part(frame) {
+  border-left-color: var(--color-yellow);
+  color: var(--color-yellow);
+}
+
+vite-error-overlay::part(stack) {
+  border-left-color: #FF5722;
+}
+
+
+vite-error-overlay::part(tip) {
+  border-top: none;
+  border-left-color: rgb(172, 127, 244);
+}
+
+vite-error-overlay::part(file):before,
+vite-error-overlay::part(frame):before,
+vite-error-overlay::part(stack):before {
+  border-bottom: 1px solid #222;
+  color: var(--color-bright);
+  display: block;
+  margin-bottom: 15px;
+  padding-bottom: 5px;
+  padding-left: 30px; /* space for icon */
+  font-size: .8rem;
+}
+
+vite-error-overlay::part(file):before {
+  content: 'File';
+}
+
+vite-error-overlay::part(frame):before {
+  content: 'Frame';
+}
+
+vite-error-overlay::part(stack):before {
+  content: 'Stack Trace';
+}
+
+vite-error-overlay::part(file):after,
+vite-error-overlay::part(frame):after,
+vite-error-overlay::part(stack):after {
+  content: '';
+  display: block;
+  height: 20px;
+  position: absolute;
+  left: var(--box-padding);
+  top: var(--box-padding);
+  width: 20px;
+}
+
+vite-error-overlay::part(file):after {
+  background-image: url("data:image/svg+xml,%3Csvg width='20px' height='20px' viewBox='0 0 24 24' version='1.1' xmlns='http://www.w3.org/2000/svg' xmlns:xlink='http://www.w3.org/1999/xlink'%3E%3Ctitle%3EFile-Generic%3C/title%3E%3Cg id='Page-1' stroke='none' stroke-width='1' fill='none' fill-rule='evenodd'%3E%3Cg id='File-Generic'%3E%3Crect id='Rectangle' fill-rule='nonzero' x='0' y='0' width='24' height='24'%3E%3C/rect%3E%3Cpath d='M4 5 C4 3.89543 4.89543 3 6 3 L15.1716 3 C15.702 3 16.2107 3.21071 16.5858 3.58579 L19.4142 6.41421 C19.7893 6.78929 20 7.29799 20 7.82843 L20 19 C20 20.1046 19.1046 21 18 21 L6 21 C4.89543 21 4 20.1046 4 19 L4 5 Z' id='Path' stroke='rgba(255,255,255,0.7)' stroke-width='1' stroke-linecap='round'%3E%3C/path%3E%3Cpath d='M15 4 L15 6 C15 7.10457 15.8954 8 17 8 L19 8' id='Path' stroke='rgba(255,255,255,0.7)' stroke-width='1' stroke-linecap='round'%3E%3C/path%3E%3C/g%3E%3C/g%3E%3C/svg%3E");
+}
+
+vite-error-overlay::part(frame):after {
+  background-image: url("data:image/svg+xml,%3Csvg width='20px' height='20px' viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M15.6602 2.84952H19.1516C20.2555 2.84952 21.1504 3.74444 21.1504 4.84839V8.3398' stroke='rgba(255,255,255,0.7)' stroke-width='1.69904' stroke-linecap='round'/%3E%3Cpath d='M2.84949 8.33981L2.84949 4.8484C2.84949 3.74446 3.74441 2.84953 4.84836 2.84953L8.33977 2.84953' stroke='rgba(255,255,255,0.7)' stroke-width='1.69904' stroke-linecap='round'/%3E%3Cpath d='M21.1505 15.6602L21.1505 19.1516C21.1505 20.2555 20.2556 21.1505 19.1516 21.1505L15.6602 21.1505' stroke='rgba(255,255,255,0.7)' stroke-width='1.69904' stroke-linecap='round'/%3E%3Cpath d='M8.33984 21.1505L4.84843 21.1505C3.74449 21.1505 2.84956 20.2555 2.84956 19.1516L2.84956 15.6602' stroke='rgba(255,255,255,0.7)' stroke-width='1.69904' stroke-linecap='round'/%3E%3C/svg%3E");
+}
+
+vite-error-overlay::part(stack):after {
+  background-image: url("data:image/svg+xml,%3Csvg width='20px' height='20px' viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M14.78 20H9.78C7.98 20 4.58 19.09 4.58 15.64C4.58 12.19 7.98 11.28 9.78 11.28H14.22C14.37 11.28 17.92 11.23 17.92 8.42C17.92 5.61 14.37 5.56 14.22 5.56H9.22C9.02109 5.56 8.83032 5.48098 8.68967 5.34033C8.54902 5.19968 8.47 5.00891 8.47 4.81C8.47 4.61109 8.54902 4.42032 8.68967 4.27967C8.83032 4.13902 9.02109 4.06 9.22 4.06H14.22C16.02 4.06 19.42 4.97 19.42 8.42C19.42 11.87 16.02 12.78 14.22 12.78H9.78C9.63 12.78 6.08 12.83 6.08 15.64C6.08 18.45 9.63 18.5 9.78 18.5H14.78C14.9789 18.5 15.1697 18.579 15.3103 18.7197C15.451 18.8603 15.53 19.0511 15.53 19.25C15.53 19.4489 15.451 19.6397 15.3103 19.7803C15.1697 19.921 14.9789 20 14.78 20Z' fill='rgba(255,255,255,0.7)'/%3E%3Cpath d='M6.44 8.31C5.74314 8.30407 5.06363 8.09202 4.48708 7.70056C3.91054 7.30909 3.46276 6.75573 3.20018 6.11021C2.93759 5.46469 2.87195 4.75589 3.01153 4.07312C3.1511 3.39036 3.48965 2.76418 3.9845 2.2735C4.47935 1.78281 5.10837 1.44958 5.79229 1.31579C6.47622 1.182 7.18444 1.25363 7.82771 1.52167C8.47099 1.78971 9.02054 2.24215 9.40711 2.82199C9.79368 3.40182 9.99998 4.08311 10 4.78C10 5.2461 9.90773 5.70759 9.72846 6.13783C9.54919 6.56808 9.28648 6.95856 8.95551 7.28675C8.62453 7.61494 8.23184 7.87433 7.80009 8.04995C7.36834 8.22558 6.90609 8.31396 6.44 8.31ZM6.44 2.75C6.04444 2.75 5.65776 2.86729 5.32886 3.08706C4.99996 3.30682 4.74362 3.61918 4.59224 3.98463C4.44087 4.35008 4.40126 4.75221 4.47843 5.14018C4.5556 5.52814 4.74609 5.8845 5.02579 6.16421C5.3055 6.44391 5.66186 6.6344 6.04982 6.71157C6.43779 6.78874 6.83992 6.74913 7.20537 6.59776C7.57082 6.44638 7.88318 6.19003 8.10294 5.86114C8.32271 5.53224 8.44 5.14556 8.44 4.75C8.44 4.48735 8.38827 4.22728 8.28776 3.98463C8.18725 3.74198 8.03993 3.5215 7.85422 3.33578C7.6685 3.15007 7.44802 3.00275 7.20537 2.90224C6.96272 2.80173 6.70265 2.75 6.44 2.75Z' fill='rgba(255,255,255,0.7)'/%3E%3Cpath d='M17.56 22.75C16.8614 22.752 16.1779 22.5466 15.5961 22.1599C15.0143 21.7733 14.5603 21.2227 14.2916 20.5778C14.0229 19.933 13.9515 19.2229 14.0866 18.5375C14.2217 17.8521 14.5571 17.2221 15.0504 16.7275C15.5437 16.2328 16.1726 15.8956 16.8577 15.7586C17.5427 15.6215 18.253 15.6909 18.8986 15.9577C19.5442 16.2246 20.0961 16.6771 20.4844 17.2578C20.8727 17.8385 21.08 18.5214 21.08 19.22C21.08 20.1545 20.7095 21.0508 20.0496 21.7125C19.3898 22.3743 18.4945 22.7473 17.56 22.75ZM17.56 17.19C17.1644 17.19 16.7778 17.3073 16.4489 17.5271C16.12 17.7468 15.8636 18.0592 15.7122 18.4246C15.5609 18.7901 15.5213 19.1922 15.5984 19.5802C15.6756 19.9681 15.8661 20.3245 16.1458 20.6042C16.4255 20.8839 16.7819 21.0744 17.1698 21.1516C17.5578 21.2287 17.9599 21.1891 18.3254 21.0377C18.6908 20.8864 19.0032 20.63 19.2229 20.3011C19.4427 19.9722 19.56 19.5856 19.56 19.19C19.56 18.6596 19.3493 18.1508 18.9742 17.7758C18.5991 17.4007 18.0904 17.19 17.56 17.19Z' fill='rgba(255,255,255,0.7)'/%3E%3C/svg%3E");
+}
+
+vite-error-overlay::part(tip):before {
+  content: "Not sure how to solve this? Visit https://qwik.builder.io or connect with the community on Discord.";
+  display: block;
+  margin-bottom: 1em;
+}
+</style>
+<script type="module" src="/@vite/client"></script>
+
+<script>
+
+document.addEventListener('qerror', ev => {
+  const ErrorOverlay = customElements.get('vite-error-overlay');
+  if (!ErrorOverlay) {
+    return;
+  }
+  const err = ev.detail.error;
+  const overlay = new ErrorOverlay(err);
+  document.body.appendChild(overlay);
+});
+</script>
+
+<script>
+/**
+ * Usage:
+ *
+ *  <errored-host></errored-host>
+ *
+ */
+class ErroredHost extends HTMLElement {
+  get _root() {
+    return this.shadowRoot || this;
+  }
+
+  constructor() {
+    super();
+    const self = this;
+
+    this.state = {};
+    if (!this.props) {
+      this.props = {};
+    }
+
+    this.componentProps = ["children", "error"];
+
+    // used to keep track of all nodes created by show/for
+    this.nodesToDestroy = [];
+    // batch updates
+    this.pendingUpdate = false;
+
+    this.attachShadow({ mode: "open" });
+  }
+
+  destroyAnyNodes() {
+    // destroy current view template refs before rendering again
+    this.nodesToDestroy.forEach((el) => el.remove());
+    this.nodesToDestroy = [];
+  }
+
+  connectedCallback() {
+    this.getAttributeNames().forEach((attr) => {
+      const jsVar = attr.replace(/-/g, "");
+      const regexp = new RegExp(jsVar, "i");
+      this.componentProps.forEach((prop) => {
+        if (regexp.test(prop)) {
+          const attrValue = this.getAttribute(attr);
+          if (this.props[prop] !== attrValue) {
+            this.props[prop] = attrValue;
+          }
+        }
+      });
+    });
+
+    this._root.innerHTML = \`
+
+      <template data-el="show-errored-host">
+      <div class="error">
+        <template data-el="div-errored-host-2">
+          <!-- String(props.error) -->
+        </template>
+      </div>
+      </template>
+
+      <div class="arrow">👇 Uncaught error happened here 👇
+        <span class="dev-tools">DevTools: Cmd+Alt+I</span>
+      </div>
+      <div class="div">
+        <slot></slot>
+      </div>
+
+      <style>
+        .error {
+          border-radius: 5px 5px 0px 0;
+          background: black;
+          color: white;
+          font-family: monospace;
+          font-size: 12px;
+          margin: 0;
+          padding: 10px;
+        }
+        .arrow {
+          background: #f47e81;
+          color: black;
+          font-size: 14px;
+          padding: 10px;
+          text-align: center;
+          font-family: sans-serif;
+        }
+        .dev-tools {
+          background: red;
+          padding: 2px 5px;
+          border-radius: 3px;
+          font-weight: 800;
+        }
+        .div {
+          outline: 5px solid red;
+          border-radius: 10px;
+        }
+      </style>\`;
+    this.pendingUpdate = true;
+
+    this.render();
+    this.onMount();
+    this.pendingUpdate = false;
+    this.update();
+  }
+
+  showContent(el) {
+    // https://developer.mozilla.org/en-US/docs/Web/API/HTMLTemplateElement/content
+    // grabs the content of a node that is between <template> tags
+    // iterates through child nodes to register all content including text elements
+    // attaches the content after the template
+
+    const elementFragment = el.content.cloneNode(true);
+    const children = Array.from(elementFragment.childNodes);
+    children.forEach((child) => {
+      if (el?.scope) {
+        child.scope = el.scope;
+      }
+      if (el?.context) {
+        child.context = el.context;
+      }
+      this.nodesToDestroy.push(child);
+    });
+    el.after(elementFragment);
+  }
+
+  onMount() {}
+
+  onUpdate() {}
+
+  update() {
+    if (this.pendingUpdate === true) {
+      return;
+    }
+    this.pendingUpdate = true;
+    this.render();
+    this.onUpdate();
+    this.pendingUpdate = false;
+  }
+
+  render() {
+    // re-rendering needs to ensure that all nodes generated by for/show are refreshed
+    this.destroyAnyNodes();
+    this.updateBindings();
+  }
+
+  updateBindings() {
+    this._root
+      .querySelectorAll("[data-el='show-errored-host']")
+      .forEach((el) => {
+        const whenCondition = this.props.error;
+        if (whenCondition) {
+          this.showContent(el);
+        }
+      });
+
+    this._root
+      .querySelectorAll("[data-el='div-errored-host-2']")
+      .forEach((el) => {
+        this.renderTextNode(el, String(this.props.error));
+      });
+  }
+
+  // Helper to render content
+  renderTextNode(el, text) {
+    const textNode = document.createTextNode(text);
+    if (el?.scope) {
+      textNode.scope = el.scope;
+    }
+    if (el?.context) {
+      textNode.context = el.context;
+    }
+    el.after(textNode);
+    this.nodesToDestroy.push(el.nextSibling);
+  }
+}
+
+customElements.define("errored-host", ErroredHost);
+</script>
+
+
+<script>
+if (!window.__qwikViteLog) {
+  window.__qwikViteLog = true;
+  console.debug("%c⭐️ Qwik Dev SSR Mode","background: #0c75d2; color: white; padding: 2px 3px; border-radius: 2px; font-size: 0.8em;","App is running in SSR development mode!\n - Additional JS is loaded by Vite for debugging and live reloading\n - Rendering performance might not be optimal\n - Delayed interactivity because prefetching is disabled\n - Vite dev bundles do not represent production output\n\nProduction build can be tested running 'npm run preview'");
+}
+</script>
+
+<style>
+#qwik-inspector-overlay {
+  position: fixed;
+  background: rgba(24, 182, 246, 0.27);
+  pointer-events: none;
+  box-sizing: border-box;
+  border: 2px solid rgba(172, 126, 244, 0.46);
+  border-radius: 4px;
+  contain: strict;
+  cursor: pointer;
+  z-index: 999999;
+}
+#qwik-inspector-info-popup {
+  position: fixed;
+  bottom: 10px;
+  right: 10px;
+  font-family: monospace;
+  background: #000000c2;
+  color: white;
+  padding: 10px 20px;
+  border-radius: 8px;
+  box-shadow: 0 20px 25px -5px rgb(0 0 0 / 34%), 0 8px 10px -6px rgb(0 0 0 / 24%);
+  backdrop-filter: blur(4px);
+  -webkit-animation: fadeOut 0.3s 3s ease-in-out forwards;
+  animation: fadeOut 0.3s 3s ease-in-out forwards;
+  z-index: 999999;
+}
+#qwik-inspector-info-popup p {
+  margin: 0px;
+}
+@-webkit-keyframes fadeOut {
+  0% {opacity: 1;}
+  100% {opacity: 0;}
+}
+
+@keyframes fadeOut {
+  0% {opacity: 1;}
+  100% {opacity: 0; visibility: hidden;}
+}
+</style>
+<script>
+(function() {
+  console.debug("%c🔍 Qwik Click-To-Source","background: #564CE0; color: white; padding: 2px 3px; border-radius: 2px; font-size: 0.8em;","Hold-press the 'Alt' key and click a component to jump directly to the source code in your IDE!");
+  window.__qwik_inspector_state = {
+    pressedKeys: new Set(),
+  };
+  const origin = 'http://local.local';
+  const srcDir = new URL("/Users/michael-personal/git-repos/Codesmith/qwik-tutorial/demo/src/", origin);
+  const body = document.body;
+  const overlay = document.createElement('div');
+  overlay.id = 'qwik-inspector-overlay';
+  overlay.setAttribute('aria-hidden', 'true');
+  body.appendChild(overlay);
+
+  document.addEventListener(
+    'keydown',
+    (event) => {
+      window.__qwik_inspector_state.pressedKeys.add(event.code);
+      updateOverlay();
+    },
+    { capture: true }
+  );
+
+  document.addEventListener(
+    'keyup',
+    (event) => {
+      window.__qwik_inspector_state.pressedKeys.delete(event.code);
+      updateOverlay();
+    },
+    { capture: true }
+  );
+
+  document.addEventListener(
+    'mouseover',
+    (event) => {
+      if (event.target && event.target instanceof HTMLElement && event.target.dataset.qwikInspector) {
+        window.__qwik_inspector_state.hoveredElement = event.target;
+      } else {
+        window.__qwik_inspector_state.hoveredElement = undefined;
+      }
+      updateOverlay();
+    },
+    { capture: true }
+  );
+
+  document.addEventListener(
+    'click',
+    (event) => {
+      if (isActive()) {
+        window.__qwik_inspector_state.pressedKeys.clear();
+        if (event.target && event.target instanceof HTMLElement) {
+          if (event.target.dataset.qwikInspector) {
+            event.preventDefault();
+            const resolvedURL = new URL(event.target.dataset.qwikInspector, srcDir);
+            body.style.setProperty('cursor', 'progress');
+            if (resolvedURL.origin === origin) {
+              const params = new URLSearchParams();
+              params.set('file', resolvedURL.pathname);
+              fetch('/__open-in-editor?' + params.toString());
+            } else {
+              location.href = resolvedURL.href;
+            }
+          }
+        }
+      }
+    },
+    { capture: true }
+  );
+
+  document.addEventListener(
+    'contextmenu',
+    (event) => {
+      if (isActive()) {
+        window.__qwik_inspector_state.pressedKeys.clear();
+        if (event.target && event.target instanceof HTMLElement) {
+          if (event.target.dataset.qwikInspector) {
+            event.preventDefault();
+          }
+        }
+      }
+    },
+    { capture: true }
+  );
+
+  function updateOverlay() {
+    const hoverElement = window.__qwik_inspector_state.hoveredElement;
+    if (hoverElement && isActive()) {
+      const rect = hoverElement.getBoundingClientRect();
+      overlay.style.setProperty('height', rect.height + 'px');
+      overlay.style.setProperty('width', rect.width + 'px');
+      overlay.style.setProperty('top', rect.top + 'px');
+      overlay.style.setProperty('left', rect.left + 'px');
+      overlay.style.setProperty('visibility', 'visible');
+      body.style.setProperty('cursor', 'pointer');
+    } else {
+      overlay.style.setProperty('height', '0px');
+      overlay.style.setProperty('width', '0px');
+      overlay.style.setProperty('visibility', 'hidden');
+      body.style.removeProperty('cursor');
+    }
+  }
+
+  function checkKeysArePressed() {
+    const activeKeys = Array.from(window.__qwik_inspector_state.pressedKeys)
+      .map((key) => key ? key.replace(/(Left|Right)$/g, '') : undefined);
+    const clickToSourceKeys = ["Alt"];
+    return clickToSourceKeys.every((key) => activeKeys.includes(key));
+  }
+
+  function isActive() {
+    return checkKeysArePressed();
+  }
+
+  window.addEventListener('resize', updateOverlay);
+  document.addEventListener('scroll', updateOverlay);
+
+})();
+</script><div id="qwik-inspector-overlay" aria-hidden="true" style="height: 0px; width: 0px; visibility: hidden;"></div>
+<div id="qwik-inspector-info-popup" aria-hidden="true">Click-to-Source: Alt<p></p></div>
+
+</body><!--/qv--><!--/qv--><!--/qv--></html>
+`
+
+const parser = new DOMParser();
+const testDOM = parser.parseFromString(testHtml, 'text/html');
+
+
+export default testDOM;
